pyproject.toml
[tool.poetry]
name = "base_project"
version = "0.1.0"
description = ""
authors = ["base <ahmed.mans20719@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.1.2"
PyYAML = "^6.0"
django-simpleui = "^2022.7.29"
djangorestframework = "^3.14.0"
django-watchman = "^1.3.0"
django-prometheus = "^2.2.0"
mysqlclient = "^2.1.1"
rest-framework-simplejwt = "^0.0.2"
django-cors-headers = "^4.1.0"
drf-spectacular = {extras = ["sidecar"], version = "^0.26.2"}
djangorestframework-simplejwt = "^5.2.2"

[tool.poetry.extras]
safety = ["click", "packaging", "pyparsing", "toml"]
bandit = ["toml"]
mysql = ["mysqlclient", "django-db-connection-pool"]

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
